api/requirements.txt
Django==2.0.1
PyMySQL==0.9.3
django-environ==0.4.5
djangorestframework==3.11.0
cryptography==2.8